+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * [UIQM-669](https://issues.folio.org/browse/UIQM-669) Display validation error messages inline.
 * [UIQM-658](https://issues.folio.org/browse/UIQM-658) Check for duplicate LCCN (010 $a) bib or authority records.
 * [UIQM-631](https://issues.folio.org/browse/UIQM-631) Create/Edit/Derive a MARC bib/authority/holdings record > Display MARC record validation errors inline.
+* [UIQM-631](https://issues.folio.org/browse/UIQM-632) Run timer on submit during backend validation, if it takes over 2s, show modal.
 
 ## [8.0.1] (https://github.com/folio-org/ui-quick-marc/tree/v8.0.1) (2024-04-18)